added language + forecast + other targets + fixed bugs
- added laguage localization
- added watch os target
- added widget target
- added icon
- added forecast backend and frontend
- fixed sun time
- added units settings

+++ MyWeather/BackEnd/WeahterModel.swift
@@ -84,7 +84,7 @@ struct WeatherModel: Decodable
     let wind: WindModel
     let clouds: CloudModel
     let sys: SysModel
-    let timezone: Int
+    let timezone: Double
     let name: String
     let rain: RainModel?
     let snow: SnowModel?
